requirements.txt
asgiref==3.8.1
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wordcloud==1.9.4
xlrd==2.0.1
